Use the request environment to calculate entry state
An additional type parameter to PNav represents the runtime environment
in which the menu is being rendered. The use case for this is an HTTP
Request object, from which the URI and the cookies may be extracted.

This object is supplied as an implicit object to the methods of
`EntryState`.